Add abstract baseSplitActivity class to better handle attachBaseContext.
Move attachBaseContext for the Activity in dyn-modules in a base abstract
Activity class that needs to be used as base.  This avoid to have to repeat the
same code for the first activity launched in a dynamic-module.

Update README with a reference to the BaseSplitActivity.

The project has been updated to work with Android Studio v3.2.